pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.45",
    "fastapi>=0.118.0",
    "nanoid>=2.0.0",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "ruff>=0.13.3",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88
lint.select = [
  "E",    # pycodestyle errors
  "F",    # pyflakes
  "I",    # isort
  "B",    # flake8-bugbear
  "UP",   # pyupgrade
  "FAST", # fastapi
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
